Check for minimum or maximum durations

The check_duration() function subsets rows of data, retaining rows that have durations that are too fast or too slow. The function is written to work with data from Qualtrics surveys.

check_duration(
  .data,
  min_duration = 10,
  max_duration = NULL,
  duration_col = "Duration (in seconds)",
  print_tibble = TRUE,
  quiet = FALSE
)

Arguments

.data
Data frame or tibble (preferably exported from Qualtrics).

min_duration
Minimum duration that is too fast in seconds.

max_duration
Maximum duration that is too slow in seconds.

duration_col
Column name for durations.

print_tibble
Logical indicating whether to print returned tibble to console.

quiet
Logical indicating whether to print message to console.

Value

An object of the same type as .data that includes the rows with fast and/or slow duration. For a function that marks these rows, use mark_duration(). For a function that excludes these rows, use exclude_duration().

Details

Default column names are set based on output from the qualtRics::fetch_survey(). By default, minimum durations of 10 seconds are checked, but either minima or maxima can be checked with the min_duration and max_duration arguments. The function outputs to console separate messages about the number of rows that are too fast or too slow.

This function returns the fast and slow rows.

See also

Examples

# Check for durations faster than 100 seconds
data(qualtrics_text)
check_duration(qualtrics_text, min_duration = 100)

#> 4 out of 100 rows took less time than the minimum duration of 100 seconds.

#> # A tibble: 4 x 16
#> # Rowwise: 
#>   StartDate           EndDate             Status     IPAddress    Progress
#>   <dttm>              <dttm>              <chr>      <chr>           <dbl>
#> 1 2020-12-11 16:59:08 2020-12-11 17:02:05 IP Address 84.56.189.0       100
#> 2 2020-12-17 15:41:52 2020-12-17 15:46:37 IP Address 15.223.0.0         13
#> 3 2020-12-17 15:41:27 2020-12-17 15:46:45 IP Address 19.127.87.0        48
#> 4 2020-12-17 15:46:46 2020-12-17 15:49:02 IP Address 21.134.217.0      100
#> # … with 11 more variables: Duration (in seconds) <dbl>, Finished <lgl>,
#> #   RecordedDate <dttm>, ResponseId <chr>, LocationLatitude <dbl>,
#> #   LocationLongitude <dbl>, UserLanguage <chr>, Browser <chr>, Version <chr>,
#> #   Operating System <chr>, Resolution <chr>

# Remove preview data first
qualtrics_text %>%
  exclude_preview() %>%
  check_duration(min_duration = 100)

#> 2 out of 100 preview rows were excluded, leaving 98 rows.

#> 4 out of 98 rows took less time than the minimum duration of 100 seconds.

#> # A tibble: 4 x 16
#> # Rowwise: 
#>   StartDate           EndDate             Status     IPAddress    Progress
#>   <dttm>              <dttm>              <chr>      <chr>           <dbl>
#> 1 2020-12-11 16:59:08 2020-12-11 17:02:05 IP Address 84.56.189.0       100
#> 2 2020-12-17 15:41:52 2020-12-17 15:46:37 IP Address 15.223.0.0         13
#> 3 2020-12-17 15:41:27 2020-12-17 15:46:45 IP Address 19.127.87.0        48
#> 4 2020-12-17 15:46:46 2020-12-17 15:49:02 IP Address 21.134.217.0      100
#> # … with 11 more variables: Duration (in seconds) <dbl>, Finished <lgl>,
#> #   RecordedDate <dttm>, ResponseId <chr>, LocationLatitude <dbl>,
#> #   LocationLongitude <dbl>, UserLanguage <chr>, Browser <chr>, Version <chr>,
#> #   Operating System <chr>, Resolution <chr>

# Check only for durations slower than 800 seconds
qualtrics_text %>%
  exclude_preview() %>%
  check_duration(max_duration = 800)

#> 2 out of 100 preview rows were excluded, leaving 98 rows.

#> 0 out of 98 rows took less time than the minimum duration of 10 seconds.

#> 2 out of 98 rows took more time than the maximum duration of 800 seconds.

#> # A tibble: 2 x 16
#> # Rowwise: 
#>   StartDate           EndDate             Status     IPAddress  Progress
#>   <dttm>              <dttm>              <chr>      <chr>         <dbl>
#> 1 2020-12-17 15:40:53 2020-12-17 15:43:25 IP Address 22.51.31.0       99
#> 2 2020-12-17 15:46:51 2020-12-17 15:51:38 IP Address 22.51.31.0      100
#> # … with 11 more variables: Duration (in seconds) <dbl>, Finished <lgl>,
#> #   RecordedDate <dttm>, ResponseId <chr>, LocationLatitude <dbl>,
#> #   LocationLongitude <dbl>, UserLanguage <chr>, Browser <chr>, Version <chr>,
#> #   Operating System <chr>, Resolution <chr>

# Do not print rows to console
qualtrics_text %>%
  exclude_preview() %>%
  check_duration(min_duration = 100, print_tibble = FALSE)

#> 2 out of 100 preview rows were excluded, leaving 98 rows.

#> 4 out of 98 rows took less time than the minimum duration of 100 seconds.

# Do not print message to console
qualtrics_text %>%
  exclude_preview() %>%
  check_duration(min_duration = 100, quiet = TRUE)

#> 2 out of 100 preview rows were excluded, leaving 98 rows.

#> # A tibble: 4 x 16
#> # Rowwise: 
#>   StartDate           EndDate             Status     IPAddress    Progress
#>   <dttm>              <dttm>              <chr>      <chr>           <dbl>
#> 1 2020-12-11 16:59:08 2020-12-11 17:02:05 IP Address 84.56.189.0       100
#> 2 2020-12-17 15:41:52 2020-12-17 15:46:37 IP Address 15.223.0.0         13
#> 3 2020-12-17 15:41:27 2020-12-17 15:46:45 IP Address 19.127.87.0        48
#> 4 2020-12-17 15:46:46 2020-12-17 15:49:02 IP Address 21.134.217.0      100
#> # … with 11 more variables: Duration (in seconds) <dbl>, Finished <lgl>,
#> #   RecordedDate <dttm>, ResponseId <chr>, LocationLatitude <dbl>,
#> #   LocationLongitude <dbl>, UserLanguage <chr>, Browser <chr>, Version <chr>,
#> #   Operating System <chr>, Resolution <chr>