Exclude IP addresses that come from a specified country.

The exclude_ip() function removes rows of data that have IP addresses in the specified country. The function is written to work with data from Qualtrics surveys.

exclude_ip(.data, id_col = "ResponseId", ...)

Arguments

.data
Data frame or tibble (preferably exported from Qualtrics).

id_col
Column name for unique row ID (e.g., participant).

...
Inherit parameters from check function.

Value

An object of the same type as .data that excludes rows with IP addresses outside of the specified country. For a function that checks these rows, use check_ip(). For a function that marks these rows, use mark_ip().

Details

Default column names are set based on output from the qualtRics::fetch_survey(). The function uses iptools::country_ranges() to assign IP addresses to specific countries using ISO 3166-1 alpha-2 country codes.

The function outputs to console a message about the number of rows with IP addresses outside of the specified country. If there are NAs for IP addresses (likely due to including preview data---see check_preview()), it will print a message alerting to the number of rows with NAs.

See also

Examples

# Exclude IP addresses outside of the US
data(qualtrics_text)
df <- exclude_ip(qualtrics_text)

#> 4 out of 100 rows with IP addresses outside of the specified country were excluded, leaving 96 rows.

# Remove preview data first
df <- qualtrics_text %>%
  exclude_preview() %>%
  exclude_ip()

#> 2 out of 100 preview rows were excluded, leaving 98 rows.

#> 4 out of 98 rows with IP addresses outside of the specified country were excluded, leaving 94 rows.

# Exclude IP addresses outside of Germany
df <- qualtrics_text %>%
  exclude_preview() %>%
  exclude_ip(country = "DE")

#> 2 out of 100 preview rows were excluded, leaving 98 rows.

#> 94 out of 98 rows with IP addresses outside of the specified country were excluded, leaving 4 rows.