Exclude unacceptable screen resolution

The exclude_resolution() function removes rows that have unacceptable screen resolution. The function is written to work with data from Qualtrics surveys.

exclude_resolution(x, id_col = "ResponseId", silent = FALSE, ...)

Arguments

x
Data frame or tibble (preferably exported from Qualtrics).

id_col
Column name for unique row ID (e.g., participant).

silent
Logical indicating whether to print message to console. Note this argument controls the exclude message not the check message.

...
Inherit parameters from check function.

Value

An object of the same type as x that excludes rows that have unacceptable screen resolutions. For a function that checks for these rows, use check_resolution(). For a function that marks these rows, use mark_resolution().

Details

Default column names are set based on output from the qualtRics::fetch_survey().

The function outputs to console a message about the number of rows with unacceptable screen resolution.

See also

Examples

# Exclude low screen resolutions
data(qualtrics_text)
df <- exclude_resolution(qualtrics_text)

#> 4 out of 100 rows with unacceptable screen resolution were excluded, leaving 96 rows.

# Remove preview data first
df <- qualtrics_text %>%
  exclude_preview() %>%
  exclude_resolution()

#> 2 out of 100 preview rows were excluded, leaving 98 rows.

#> 4 out of 98 rows with unacceptable screen resolution were excluded, leaving 94 rows.