Mark IP addresses from outside of a specified country.

The mark_ip() function creates a column labeling rows of data that have IP addresses from outside the specified country. The function is written to work with data from Qualtrics surveys.

mark_ip(
  x,
  id_col = "ResponseId",
  ip_col = "IPAddress",
  country = "US",
  quiet = FALSE
)

Arguments

x
Data frame or tibble (preferably imported from Qualtrics using {qualtRics}).

id_col
Column name for unique row ID (e.g., participant).

ip_col
Column name for IP addresses.

country
Two-letter abbreviation of country to check (default is "US").

quiet
Logical indicating whether to print message to console.

Value

An object of the same type as x that includes a column marking rows with IP addresses outside of the specified country. For a function that checks these rows, use check_ip(). For a function that excludes these rows, use exclude_ip().

Details

Default column names are set based on output from the qualtRics::fetch_survey(). The function uses iptools::country_ranges() to assign IP addresses to specific countries using ISO 3166-1 alpha-2 country codes.

The function outputs to console a message about the number of rows with IP addresses outside of the specified country. If there are NAs for IP addresses (likely due to including preview data---see check_preview()), it will print a message alerting to the number of rows with NAs.

See also

Examples

# Mark IP addresses outside of the US
data(qualtrics_text)
df <- mark_ip(qualtrics_text)

#> 2 out of 100 rows have NA values for IP addresses (check for preview data with 'check_preview()').

#> 4 out of 100 rows have IP addresses outside of US.

# Remove preview data first
df <- qualtrics_text %>%
  exclude_preview() %>%
  mark_ip()

#> 2 out of 100 duplicate rows were excluded, leaving 98 rows.

#> 4 out of 98 rows have IP addresses outside of US.

# Mark IP addresses outside of Germany
df <- qualtrics_text %>%
  exclude_preview() %>%
  mark_ip(country = "DE")

#> 2 out of 100 duplicate rows were excluded, leaving 98 rows.

#> 94 out of 98 rows have IP addresses outside of DE.