Mark survey previews

The mark_preview() function creates a column labeling rows that are survey previews. The function is written to work with data from Qualtrics surveys.

mark_preview(.data, id_col = "ResponseId", ...)

Arguments

.data
Data frame or tibble (preferably exported from Qualtrics).

id_col
Column name for unique row ID (e.g., participant).

...
Inherit parameters from check function.

Value

An object of the same type as .data that includes a column marking rows that are survey previews. For a function that checks for these rows, use check_preview(). For a function that excludes these rows, use exclude_preview().

Details

Default column names are set based on output from the qualtRics::fetch_survey(). The preview column in Qualtrics can be a numeric or character vector depending on whether it is exported as choice text or numeric values. This function works for both.

The function outputs to console a message about the number of rows that are survey previews.

See also

Examples

# Mark survey previews
data(qualtrics_text)
df <- mark_preview(qualtrics_text)

#> 2 out of 100 rows were collected as previews. It is highly recommended to exclude these rows before further checking.

# Works for Qualtrics data exported as numeric values, too
df <- qualtrics_numeric %>%
  mark_preview()

#> 2 out of 100 rows were collected as previews. It is highly recommended to exclude these rows before further checking.

# Do not print rows to console
df <- qualtrics_text %>%
  mark_preview(print_tibble = FALSE)

#> 2 out of 100 rows were collected as previews. It is highly recommended to exclude these rows before further checking.