Check for duplicate IP addresses and/or locations

The check_duplicates() function subsets rows of data, retaining rows that have the same IP address and/or same latitude and longitude. The function is written to work with data from Qualtrics surveys.

check_duplicates(
  .data,
  ip_col = "IPAddress",
  location_col = c("LocationLatitude", "LocationLongitude"),
  dupl_ip = TRUE,
  dupl_location = TRUE,
  include_na = FALSE,
  print_tibble = TRUE,
  quiet = FALSE
)

Arguments

.data
Data frame or tibble (preferably exported from Qualtrics).

ip_col
Column name for IP addresses.

location_col
Two element vector specifying columns for latitude and longitude (in that order).

dupl_ip
Logical indicating whether to check IP addresses.

dupl_location
Logical indicating whether to check latitude and longitude.

include_na
Logical indicating whether to include rows with NAs for IP address and location as potentially excluded rows.

print_tibble
Logical indicating whether to print returned tibble to console.

quiet
Logical indicating whether to print message to console.

Value

An object of the same type as .data that includes the rows with duplicate IP addresses and/or locations. For a function that marks these rows, use mark_duplicates(). For a function that excludes these rows, use exclude_duplicates().

Details

Default column names are set based on output from the qualtRics::fetch_survey(). By default, IP address and location are both checked, but they can be checked separately with the dupl_ip and dupl_location arguments.

The function outputs to console separate messages about the number of rows with duplicate IP addresses and rows with duplicate locations. These counts are computed independently, so rows may be counted for both types of duplicates.

See also

Examples

# Check for duplicate IP addresses and locations
data(qualtrics_text)
check_duplicates(qualtrics_text)

#> 

#> 10 out of 99 rows have duplicate locations.

#> 7 out of 97 rows have duplicate IP addresses.

#> # A tibble: 17 x 17
#>    LocationLatitude LocationLongitu… dupe_count StartDate          
#>               <dbl>            <dbl>      <int> <dttm>             
#>  1             28.6            -81.5          2 2020-12-17 15:42:47
#>  2             28.6            -81.5          2 2020-12-17 15:42:18
#>  3             37.3           -121.           3 2020-12-17 15:40:53
#>  4             37.3           -121.           3 2020-12-17 15:46:51
#>  5             37.3           -121.           3 2020-12-17 15:48:53
#>  6             40.3            -75.9          2 2020-12-11 12:41:23
#>  7             40.3            -75.9          2 2020-12-17 15:41:17
#>  8             45.5           -123.           3 2020-12-17 15:40:52
#>  9             45.5           -123.           3 2020-12-17 15:40:57
#> 10             45.5           -123.           3 2020-12-17 15:48:48
#> 11             37.3           -121.           3 2020-12-17 15:40:53
#> 12             37.3           -121.           3 2020-12-17 15:46:51
#> 13             37.3           -121.           3 2020-12-17 15:48:53
#> 14             40.3            -75.9          2 2020-12-11 12:41:23
#> 15             40.3            -75.9          2 2020-12-17 15:41:17
#> 16             28.6            -81.5          2 2020-12-17 15:42:47
#> 17             28.6            -81.5          2 2020-12-17 15:42:18
#> # … with 13 more variables: EndDate <dttm>, Status <chr>, IPAddress <chr>,
#> #   Progress <dbl>, `Duration (in seconds)` <dbl>, Finished <lgl>,
#> #   RecordedDate <dttm>, ResponseId <chr>, UserLanguage <chr>, Browser <chr>,
#> #   Version <chr>, `Operating System` <chr>, Resolution <chr>

# Check only for duplicate locations
qualtrics_text %>%
  check_duplicates(dupl_location = FALSE)

#> 7 out of 98 rows have duplicate IP addresses.

#> # A tibble: 7 x 17
#>   IPAddress dupe_count StartDate           EndDate             Status Progress
#>   <chr>          <int> <dttm>              <dttm>              <chr>     <dbl>
#> 1 22.51.31…          3 2020-12-17 15:40:53 2020-12-17 15:43:25 IP Ad…       99
#> 2 22.51.31…          3 2020-12-17 15:46:51 2020-12-17 15:51:38 IP Ad…      100
#> 3 22.51.31…          3 2020-12-17 15:48:53 2020-12-17 15:53:48 IP Ad…      100
#> 4 24.195.9…          2 2020-12-11 12:41:23 2020-12-11 12:44:37 IP Ad…      100
#> 5 24.195.9…          2 2020-12-17 15:41:17 2020-12-17 15:45:42 IP Ad…      100
#> 6 55.73.11…          2 2020-12-17 15:42:47 2020-12-17 15:46:26 IP Ad…      100
#> 7 55.73.11…          2 2020-12-17 15:42:18 2020-12-17 15:48:00 IP Ad…      100
#> # … with 11 more variables: `Duration (in seconds)` <dbl>, Finished <lgl>,
#> #   RecordedDate <dttm>, ResponseId <chr>, LocationLatitude <dbl>,
#> #   LocationLongitude <dbl>, UserLanguage <chr>, Browser <chr>, Version <chr>,
#> #   `Operating System` <chr>, Resolution <chr>

# Do not print rows to console
qualtrics_text %>%
  check_duplicates(print_tibble = FALSE)

#> 

#> 10 out of 99 rows have duplicate locations.

#> 7 out of 97 rows have duplicate IP addresses.

# Do not print message to console
qualtrics_text %>%
  check_duplicates(quiet = TRUE)

#> # A tibble: 17 x 17
#>    LocationLatitude LocationLongitu… dupe_count StartDate          
#>               <dbl>            <dbl>      <int> <dttm>             
#>  1             28.6            -81.5          2 2020-12-17 15:42:47
#>  2             28.6            -81.5          2 2020-12-17 15:42:18
#>  3             37.3           -121.           3 2020-12-17 15:40:53
#>  4             37.3           -121.           3 2020-12-17 15:46:51
#>  5             37.3           -121.           3 2020-12-17 15:48:53
#>  6             40.3            -75.9          2 2020-12-11 12:41:23
#>  7             40.3            -75.9          2 2020-12-17 15:41:17
#>  8             45.5           -123.           3 2020-12-17 15:40:52
#>  9             45.5           -123.           3 2020-12-17 15:40:57
#> 10             45.5           -123.           3 2020-12-17 15:48:48
#> 11             37.3           -121.           3 2020-12-17 15:40:53
#> 12             37.3           -121.           3 2020-12-17 15:46:51
#> 13             37.3           -121.           3 2020-12-17 15:48:53
#> 14             40.3            -75.9          2 2020-12-11 12:41:23
#> 15             40.3            -75.9          2 2020-12-17 15:41:17
#> 16             28.6            -81.5          2 2020-12-17 15:42:47
#> 17             28.6            -81.5          2 2020-12-17 15:42:18
#> # … with 13 more variables: EndDate <dttm>, Status <chr>, IPAddress <chr>,
#> #   Progress <dbl>, `Duration (in seconds)` <dbl>, Finished <lgl>,
#> #   RecordedDate <dttm>, ResponseId <chr>, UserLanguage <chr>, Browser <chr>,
#> #   Version <chr>, `Operating System` <chr>, Resolution <chr>