Check for duplicate IP addresses and/or locations

The check_duplicates() function subsets rows of data, retaining rows that have the same IP address and/or same latitude and longitude. The function is written to work with data from Qualtrics surveys.

check_duplicates(
  x,
  ip_col = "IPAddress",
  location_col = c("LocationLatitude", "LocationLongitude"),
  dupl_ip = TRUE,
  dupl_location = TRUE,
  include_na = FALSE,
  print_tibble = TRUE,
  quiet = FALSE
)

Arguments

x
Data frame or tibble (preferably exported from Qualtrics).

ip_col
Column name for IP addresses.

location_col
Two element vector specifying columns for latitude and longitude (in that order).

dupl_ip
Logical indicating whether to check IP addresses.

dupl_location
Logical indicating whether to check latitude and longitude.

include_na
Logical indicating whether to include rows with NAs for IP address and location as potentially excluded rows.

print_tibble
Logical indicating whether to print returned tibble to console.

quiet
Logical indicating whether to print message to console.x

Value

An object of the same type as x that includes the rows with duplicate IP addresses and/or locations. This includes a column called dupe_count that returns the number of duplicates. For a function that marks these rows, use mark_duplicates(). For a function that excludes these rows, use exclude_duplicates().

Details

Default column names are set based on output from the qualtRics::fetch_survey(). By default, IP address and location are both checked, but they can be checked separately with the dupl_ip and dupl_location arguments.

The function outputs to console separate messages about the number of rows with duplicate IP addresses and rows with duplicate locations. These counts are computed independently, so rows may be counted for both types of duplicates.

See also

Examples

# Check for duplicate IP addresses and locations
data(qualtrics_text)
check_duplicates(qualtrics_text)

#> 7 out of 98 rows have duplicate IP addresses.

#> 0 NAs were found in location.

#> 10 out of 97 rows have duplicate locations.

#>    LocationLatitude LocationLongitude           StartDate             EndDate
#> 1          28.56411         -81.54902 2020-12-17 15:42:47 2020-12-17 15:46:26
#> 2          28.56411         -81.54902 2020-12-17 15:42:18 2020-12-17 15:48:00
#> 3          37.28265        -120.50248 2020-12-17 15:40:53 2020-12-17 15:43:25
#> 4          37.28265        -120.50248 2020-12-17 15:46:51 2020-12-17 15:51:38
#> 5          37.28265        -120.50248 2020-12-17 15:48:53 2020-12-17 15:53:48
#> 6          40.33554         -75.92698 2020-12-11 12:41:23 2020-12-11 12:44:37
#> 7          40.33554         -75.92698 2020-12-17 15:41:17 2020-12-17 15:45:42
#> 8          45.50412        -122.78665 2020-12-17 15:40:52 2020-12-17 15:43:39
#> 9          45.50412        -122.78665 2020-12-17 15:40:57 2020-12-17 15:48:56
#> 10         45.50412        -122.78665 2020-12-17 15:48:48 2020-12-17 15:54:12
#> 11         37.28265        -120.50248 2020-12-17 15:40:53 2020-12-17 15:43:25
#> 12         37.28265        -120.50248 2020-12-17 15:46:51 2020-12-17 15:51:38
#> 13         37.28265        -120.50248 2020-12-17 15:48:53 2020-12-17 15:53:48
#> 14         40.33554         -75.92698 2020-12-11 12:41:23 2020-12-11 12:44:37
#> 15         40.33554         -75.92698 2020-12-17 15:41:17 2020-12-17 15:45:42
#> 16         28.56411         -81.54902 2020-12-17 15:42:47 2020-12-17 15:46:26
#> 17         28.56411         -81.54902 2020-12-17 15:42:18 2020-12-17 15:48:00
#>        Status    IPAddress Progress Duration (in seconds) Finished
#> 1  IP Address  55.73.114.0      100                   236     TRUE
#> 2  IP Address  55.73.114.0      100                   526     TRUE
#> 3  IP Address   22.51.31.0       99                   879    FALSE
#> 4  IP Address   22.51.31.0      100                   872     TRUE
#> 5  IP Address   22.51.31.0      100                   246     TRUE
#> 6  IP Address  24.195.91.0      100                   177     TRUE
#> 7  IP Address  24.195.91.0      100                   521     TRUE
#> 8  IP Address 32.164.134.0      100                   375     TRUE
#> 9  IP Address   6.79.107.0      100                   397     TRUE
#> 10 IP Address 54.232.129.0      100                   149     TRUE
#> 11 IP Address   22.51.31.0       99                   879    FALSE
#> 12 IP Address   22.51.31.0      100                   872     TRUE
#> 13 IP Address   22.51.31.0      100                   246     TRUE
#> 14 IP Address  24.195.91.0      100                   177     TRUE
#> 15 IP Address  24.195.91.0      100                   521     TRUE
#> 16 IP Address  55.73.114.0      100                   236     TRUE
#> 17 IP Address  55.73.114.0      100                   526     TRUE
#>           RecordedDate        ResponseId UserLanguage Browser       Version
#> 1  2020-12-17 15:46:26 R_7UzegytocfkyrWC           EN  Chrome  87.0.4280.88
#> 2  2020-12-17 15:48:00 R_NiK6d3RgjuJh1OI           EN  Chrome  87.0.4280.88
#> 3  2020-12-17 15:43:25 R_AkQyJypPyjgribz           EN  Chrome  87.0.4280.88
#> 4  2020-12-17 15:51:39 R_Gbz5en48KgnCXT7           EN Firefox          83.0
#> 5  2020-12-17 15:53:48 R_AJfrQqClQNvWIch           EN    Edge   84.0.522.52
#> 6  2020-12-11 12:44:37 R_LAt58JGEyKNWZlB           EN  Chrome 86.0.4240.198
#> 7  2020-12-17 15:45:42 R_GNVaLC9Sb2ZDzQP           EN  Chrome  87.0.4280.88
#> 8  2020-12-17 15:43:39 R_H5MqcQoWznreNBt           EN  Chrome  87.0.4280.88
#> 9  2020-12-17 15:48:56 R_8ezIj0X0p2lJuCQ           EN  Chrome  87.0.4280.88
#> 10 2020-12-17 15:54:12 R_Kc9BGXO793zEqHM           EN  Chrome  87.0.4280.88
#> 11 2020-12-17 15:43:25 R_AkQyJypPyjgribz           EN  Chrome  87.0.4280.88
#> 12 2020-12-17 15:51:39 R_Gbz5en48KgnCXT7           EN Firefox          83.0
#> 13 2020-12-17 15:53:48 R_AJfrQqClQNvWIch           EN    Edge   84.0.522.52
#> 14 2020-12-11 12:44:37 R_LAt58JGEyKNWZlB           EN  Chrome 86.0.4240.198
#> 15 2020-12-17 15:45:42 R_GNVaLC9Sb2ZDzQP           EN  Chrome  87.0.4280.88
#> 16 2020-12-17 15:46:26 R_7UzegytocfkyrWC           EN  Chrome  87.0.4280.88
#> 17 2020-12-17 15:48:00 R_NiK6d3RgjuJh1OI           EN  Chrome  87.0.4280.88
#>    Operating System Resolution
#> 1   Windows NT 10.0  1920x1080
#> 2   Windows NT 10.0   1536x864
#> 3   Windows NT 10.0   1366x768
#> 4   Windows NT 10.0   1440x960
#> 5   Windows NT 10.0  1920x1080
#> 6         Macintosh   1280x800
#> 7    Windows NT 6.1   1366x768
#> 8   Windows NT 10.0  1920x1080
#> 9   Windows NT 10.0   1536x864
#> 10  Windows NT 10.0  1920x1080
#> 11  Windows NT 10.0   1366x768
#> 12  Windows NT 10.0   1440x960
#> 13  Windows NT 10.0  1920x1080
#> 14        Macintosh   1280x800
#> 15   Windows NT 6.1   1366x768
#> 16  Windows NT 10.0  1920x1080
#> 17  Windows NT 10.0   1536x864

# Check only for duplicate locations
qualtrics_text %>%
  check_duplicates(dupl_location = FALSE)

#> 7 out of 98 rows have duplicate IP addresses.

#>     IPAddress           StartDate             EndDate     Status Progress
#> 1  22.51.31.0 2020-12-17 15:40:53 2020-12-17 15:43:25 IP Address       99
#> 2  22.51.31.0 2020-12-17 15:46:51 2020-12-17 15:51:38 IP Address      100
#> 3  22.51.31.0 2020-12-17 15:48:53 2020-12-17 15:53:48 IP Address      100
#> 4 24.195.91.0 2020-12-11 12:41:23 2020-12-11 12:44:37 IP Address      100
#> 5 24.195.91.0 2020-12-17 15:41:17 2020-12-17 15:45:42 IP Address      100
#> 6 55.73.114.0 2020-12-17 15:42:47 2020-12-17 15:46:26 IP Address      100
#> 7 55.73.114.0 2020-12-17 15:42:18 2020-12-17 15:48:00 IP Address      100
#>   Duration (in seconds) Finished        RecordedDate        ResponseId
#> 1                   879    FALSE 2020-12-17 15:43:25 R_AkQyJypPyjgribz
#> 2                   872     TRUE 2020-12-17 15:51:39 R_Gbz5en48KgnCXT7
#> 3                   246     TRUE 2020-12-17 15:53:48 R_AJfrQqClQNvWIch
#> 4                   177     TRUE 2020-12-11 12:44:37 R_LAt58JGEyKNWZlB
#> 5                   521     TRUE 2020-12-17 15:45:42 R_GNVaLC9Sb2ZDzQP
#> 6                   236     TRUE 2020-12-17 15:46:26 R_7UzegytocfkyrWC
#> 7                   526     TRUE 2020-12-17 15:48:00 R_NiK6d3RgjuJh1OI
#>   LocationLatitude LocationLongitude UserLanguage Browser       Version
#> 1         37.28265        -120.50248           EN  Chrome  87.0.4280.88
#> 2         37.28265        -120.50248           EN Firefox          83.0
#> 3         37.28265        -120.50248           EN    Edge   84.0.522.52
#> 4         40.33554         -75.92698           EN  Chrome 86.0.4240.198
#> 5         40.33554         -75.92698           EN  Chrome  87.0.4280.88
#> 6         28.56411         -81.54902           EN  Chrome  87.0.4280.88
#> 7         28.56411         -81.54902           EN  Chrome  87.0.4280.88
#>   Operating System Resolution
#> 1  Windows NT 10.0   1366x768
#> 2  Windows NT 10.0   1440x960
#> 3  Windows NT 10.0  1920x1080
#> 4        Macintosh   1280x800
#> 5   Windows NT 6.1   1366x768
#> 6  Windows NT 10.0  1920x1080
#> 7  Windows NT 10.0   1536x864

# Do not print rows to console
qualtrics_text %>%
  check_duplicates(print_tibble = FALSE)

#> 7 out of 98 rows have duplicate IP addresses.

#> 0 NAs were found in location.

#> 10 out of 97 rows have duplicate locations.

# Do not print message to console
qualtrics_text %>%
  check_duplicates(quiet = TRUE)

#>    LocationLatitude LocationLongitude           StartDate             EndDate
#> 1          28.56411         -81.54902 2020-12-17 15:42:47 2020-12-17 15:46:26
#> 2          28.56411         -81.54902 2020-12-17 15:42:18 2020-12-17 15:48:00
#> 3          37.28265        -120.50248 2020-12-17 15:40:53 2020-12-17 15:43:25
#> 4          37.28265        -120.50248 2020-12-17 15:46:51 2020-12-17 15:51:38
#> 5          37.28265        -120.50248 2020-12-17 15:48:53 2020-12-17 15:53:48
#> 6          40.33554         -75.92698 2020-12-11 12:41:23 2020-12-11 12:44:37
#> 7          40.33554         -75.92698 2020-12-17 15:41:17 2020-12-17 15:45:42
#> 8          45.50412        -122.78665 2020-12-17 15:40:52 2020-12-17 15:43:39
#> 9          45.50412        -122.78665 2020-12-17 15:40:57 2020-12-17 15:48:56
#> 10         45.50412        -122.78665 2020-12-17 15:48:48 2020-12-17 15:54:12
#> 11         37.28265        -120.50248 2020-12-17 15:40:53 2020-12-17 15:43:25
#> 12         37.28265        -120.50248 2020-12-17 15:46:51 2020-12-17 15:51:38
#> 13         37.28265        -120.50248 2020-12-17 15:48:53 2020-12-17 15:53:48
#> 14         40.33554         -75.92698 2020-12-11 12:41:23 2020-12-11 12:44:37
#> 15         40.33554         -75.92698 2020-12-17 15:41:17 2020-12-17 15:45:42
#> 16         28.56411         -81.54902 2020-12-17 15:42:47 2020-12-17 15:46:26
#> 17         28.56411         -81.54902 2020-12-17 15:42:18 2020-12-17 15:48:00
#>        Status    IPAddress Progress Duration (in seconds) Finished
#> 1  IP Address  55.73.114.0      100                   236     TRUE
#> 2  IP Address  55.73.114.0      100                   526     TRUE
#> 3  IP Address   22.51.31.0       99                   879    FALSE
#> 4  IP Address   22.51.31.0      100                   872     TRUE
#> 5  IP Address   22.51.31.0      100                   246     TRUE
#> 6  IP Address  24.195.91.0      100                   177     TRUE
#> 7  IP Address  24.195.91.0      100                   521     TRUE
#> 8  IP Address 32.164.134.0      100                   375     TRUE
#> 9  IP Address   6.79.107.0      100                   397     TRUE
#> 10 IP Address 54.232.129.0      100                   149     TRUE
#> 11 IP Address   22.51.31.0       99                   879    FALSE
#> 12 IP Address   22.51.31.0      100                   872     TRUE
#> 13 IP Address   22.51.31.0      100                   246     TRUE
#> 14 IP Address  24.195.91.0      100                   177     TRUE
#> 15 IP Address  24.195.91.0      100                   521     TRUE
#> 16 IP Address  55.73.114.0      100                   236     TRUE
#> 17 IP Address  55.73.114.0      100                   526     TRUE
#>           RecordedDate        ResponseId UserLanguage Browser       Version
#> 1  2020-12-17 15:46:26 R_7UzegytocfkyrWC           EN  Chrome  87.0.4280.88
#> 2  2020-12-17 15:48:00 R_NiK6d3RgjuJh1OI           EN  Chrome  87.0.4280.88
#> 3  2020-12-17 15:43:25 R_AkQyJypPyjgribz           EN  Chrome  87.0.4280.88
#> 4  2020-12-17 15:51:39 R_Gbz5en48KgnCXT7           EN Firefox          83.0
#> 5  2020-12-17 15:53:48 R_AJfrQqClQNvWIch           EN    Edge   84.0.522.52
#> 6  2020-12-11 12:44:37 R_LAt58JGEyKNWZlB           EN  Chrome 86.0.4240.198
#> 7  2020-12-17 15:45:42 R_GNVaLC9Sb2ZDzQP           EN  Chrome  87.0.4280.88
#> 8  2020-12-17 15:43:39 R_H5MqcQoWznreNBt           EN  Chrome  87.0.4280.88
#> 9  2020-12-17 15:48:56 R_8ezIj0X0p2lJuCQ           EN  Chrome  87.0.4280.88
#> 10 2020-12-17 15:54:12 R_Kc9BGXO793zEqHM           EN  Chrome  87.0.4280.88
#> 11 2020-12-17 15:43:25 R_AkQyJypPyjgribz           EN  Chrome  87.0.4280.88
#> 12 2020-12-17 15:51:39 R_Gbz5en48KgnCXT7           EN Firefox          83.0
#> 13 2020-12-17 15:53:48 R_AJfrQqClQNvWIch           EN    Edge   84.0.522.52
#> 14 2020-12-11 12:44:37 R_LAt58JGEyKNWZlB           EN  Chrome 86.0.4240.198
#> 15 2020-12-17 15:45:42 R_GNVaLC9Sb2ZDzQP           EN  Chrome  87.0.4280.88
#> 16 2020-12-17 15:46:26 R_7UzegytocfkyrWC           EN  Chrome  87.0.4280.88
#> 17 2020-12-17 15:48:00 R_NiK6d3RgjuJh1OI           EN  Chrome  87.0.4280.88
#>    Operating System Resolution
#> 1   Windows NT 10.0  1920x1080
#> 2   Windows NT 10.0   1536x864
#> 3   Windows NT 10.0   1366x768
#> 4   Windows NT 10.0   1440x960
#> 5   Windows NT 10.0  1920x1080
#> 6         Macintosh   1280x800
#> 7    Windows NT 6.1   1366x768
#> 8   Windows NT 10.0  1920x1080
#> 9   Windows NT 10.0   1536x864
#> 10  Windows NT 10.0  1920x1080
#> 11  Windows NT 10.0   1366x768
#> 12  Windows NT 10.0   1440x960
#> 13  Windows NT 10.0  1920x1080
#> 14        Macintosh   1280x800
#> 15   Windows NT 6.1   1366x768
#> 16  Windows NT 10.0  1920x1080
#> 17  Windows NT 10.0   1536x864