Check for minimum or maximum durations

The check_duration() function subsets rows of data, retaining rows that have durations that are too fast or too slow. The function is written to work with data from Qualtrics surveys.

check_duration(
  x,
  min_duration = 10,
  max_duration = NULL,
  duration_col = "Duration (in seconds)",
  print_tibble = TRUE,
  quiet = FALSE
)

Arguments

x
Data frame or tibble (preferably exported from Qualtrics).

min_duration
Minimum duration that is too fast in seconds.

max_duration
Maximum duration that is too slow in seconds.

duration_col
Column name for durations.

print_tibble
Logical indicating whether to print returned tibble to console.

quiet
Logical indicating whether to print message to console.

Value

An object of the same type as x that includes the rows with fast and/or slow duration. For a function that marks these rows, use mark_duration(). For a function that excludes these rows, use exclude_duration().

Details

Default column names are set based on output from the qualtRics::fetch_survey(). By default, minimum durations of 10 seconds are checked, but either minima or maxima can be checked with the min_duration and max_duration arguments. The function outputs to console separate messages about the number of rows that are too fast or too slow.

This function returns the fast and slow rows.

See also

Examples

# Check for durations faster than 100 seconds
data(qualtrics_text)
check_duration(qualtrics_text, min_duration = 100)

#> 4 out of 100 rows took less time than the minimum duration of 100 seconds.

#>              StartDate             EndDate     Status    IPAddress Progress
#> 11 2020-12-11 16:59:08 2020-12-11 17:02:05 IP Address  84.56.189.0      100
#> 42 2020-12-17 15:41:52 2020-12-17 15:46:37 IP Address   15.223.0.0       13
#> 44 2020-12-17 15:41:27 2020-12-17 15:46:45 IP Address  19.127.87.0       48
#> 61 2020-12-17 15:46:46 2020-12-17 15:49:02 IP Address 21.134.217.0      100
#>    Duration (in seconds) Finished        RecordedDate        ResponseId
#> 11                    54     TRUE 2020-12-11 17:02:05 R_2RQ5kfCKKHudpj3
#> 42                    40    FALSE 2020-12-17 15:46:37 R_Dx6w74UfhnGhAmj
#> 44                    74    FALSE 2020-12-17 15:46:46 R_ewyyOOPADLGo9xZ
#> 61                    72     TRUE 2020-12-17 15:49:02 R_PKKUJ04DtpTEire
#>    LocationLatitude LocationLongitude UserLanguage Browser      Version
#> 11         43.23353         -77.55959           EN  Chrome 87.0.4280.88
#> 42         34.77804         -84.96198           DE  Chrome 83.0.4103.61
#> 44         33.66715        -117.82543           EN  Chrome 87.0.4280.88
#> 61         34.76243         -96.69044           EN  Chrome 87.0.4280.88
#>    Operating System Resolution
#> 11  Windows NT 10.0   1600x900
#> 42  Windows NT 10.0   1366x768
#> 44  Windows NT 10.0   1188x668
#> 61  Windows NT 10.0   1368x912

# Remove preview data first
qualtrics_text %>%
  exclude_preview() %>%
  check_duration(min_duration = 100)

#> 2 out of 100 preview rows were excluded, leaving 98 rows.

#> 4 out of 98 rows took less time than the minimum duration of 100 seconds.

#>              StartDate             EndDate     Status    IPAddress Progress
#> 9  2020-12-11 16:59:08 2020-12-11 17:02:05 IP Address  84.56.189.0      100
#> 40 2020-12-17 15:41:52 2020-12-17 15:46:37 IP Address   15.223.0.0       13
#> 42 2020-12-17 15:41:27 2020-12-17 15:46:45 IP Address  19.127.87.0       48
#> 59 2020-12-17 15:46:46 2020-12-17 15:49:02 IP Address 21.134.217.0      100
#>    Duration (in seconds) Finished        RecordedDate        ResponseId
#> 9                     54     TRUE 2020-12-11 17:02:05 R_2RQ5kfCKKHudpj3
#> 40                    40    FALSE 2020-12-17 15:46:37 R_Dx6w74UfhnGhAmj
#> 42                    74    FALSE 2020-12-17 15:46:46 R_ewyyOOPADLGo9xZ
#> 59                    72     TRUE 2020-12-17 15:49:02 R_PKKUJ04DtpTEire
#>    LocationLatitude LocationLongitude UserLanguage Browser      Version
#> 9          43.23353         -77.55959           EN  Chrome 87.0.4280.88
#> 40         34.77804         -84.96198           DE  Chrome 83.0.4103.61
#> 42         33.66715        -117.82543           EN  Chrome 87.0.4280.88
#> 59         34.76243         -96.69044           EN  Chrome 87.0.4280.88
#>    Operating System Resolution
#> 9   Windows NT 10.0   1600x900
#> 40  Windows NT 10.0   1366x768
#> 42  Windows NT 10.0   1188x668
#> 59  Windows NT 10.0   1368x912

# Check only for durations slower than 800 seconds
qualtrics_text %>%
  exclude_preview() %>%
  check_duration(max_duration = 800)

#> 2 out of 100 preview rows were excluded, leaving 98 rows.

#> 0 out of 98 rows took less time than the minimum duration of 10 seconds.

#> 2 out of 98 rows took more time than the maximum duration of 800 seconds.

#>              StartDate             EndDate     Status  IPAddress Progress
#> 21 2020-12-17 15:40:53 2020-12-17 15:43:25 IP Address 22.51.31.0       99
#> 75 2020-12-17 15:46:51 2020-12-17 15:51:38 IP Address 22.51.31.0      100
#>    Duration (in seconds) Finished        RecordedDate        ResponseId
#> 21                   879    FALSE 2020-12-17 15:43:25 R_AkQyJypPyjgribz
#> 75                   872     TRUE 2020-12-17 15:51:39 R_Gbz5en48KgnCXT7
#>    LocationLatitude LocationLongitude UserLanguage Browser      Version
#> 21         37.28265         -120.5025           EN  Chrome 87.0.4280.88
#> 75         37.28265         -120.5025           EN Firefox         83.0
#>    Operating System Resolution
#> 21  Windows NT 10.0   1366x768
#> 75  Windows NT 10.0   1440x960

# Do not print rows to console
qualtrics_text %>%
  exclude_preview() %>%
  check_duration(min_duration = 100, print_tibble = FALSE)

#> 2 out of 100 preview rows were excluded, leaving 98 rows.

#> 4 out of 98 rows took less time than the minimum duration of 100 seconds.

# Do not print message to console
qualtrics_text %>%
  exclude_preview() %>%
  check_duration(min_duration = 100, quiet = TRUE)

#> 2 out of 100 preview rows were excluded, leaving 98 rows.

#>              StartDate             EndDate     Status    IPAddress Progress
#> 9  2020-12-11 16:59:08 2020-12-11 17:02:05 IP Address  84.56.189.0      100
#> 40 2020-12-17 15:41:52 2020-12-17 15:46:37 IP Address   15.223.0.0       13
#> 42 2020-12-17 15:41:27 2020-12-17 15:46:45 IP Address  19.127.87.0       48
#> 59 2020-12-17 15:46:46 2020-12-17 15:49:02 IP Address 21.134.217.0      100
#>    Duration (in seconds) Finished        RecordedDate        ResponseId
#> 9                     54     TRUE 2020-12-11 17:02:05 R_2RQ5kfCKKHudpj3
#> 40                    40    FALSE 2020-12-17 15:46:37 R_Dx6w74UfhnGhAmj
#> 42                    74    FALSE 2020-12-17 15:46:46 R_ewyyOOPADLGo9xZ
#> 59                    72     TRUE 2020-12-17 15:49:02 R_PKKUJ04DtpTEire
#>    LocationLatitude LocationLongitude UserLanguage Browser      Version
#> 9          43.23353         -77.55959           EN  Chrome 87.0.4280.88
#> 40         34.77804         -84.96198           DE  Chrome 83.0.4103.61
#> 42         33.66715        -117.82543           EN  Chrome 87.0.4280.88
#> 59         34.76243         -96.69044           EN  Chrome 87.0.4280.88
#>    Operating System Resolution
#> 9   Windows NT 10.0   1600x900
#> 40  Windows NT 10.0   1366x768
#> 42  Windows NT 10.0   1188x668
#> 59  Windows NT 10.0   1368x912