Check for IP addresses that come from a specified country.

The check_ip() function subsets rows of data, retaining rows that have IP addresses in the specified country. The function is written to work with data from Qualtrics surveys.

check_ip(
  .data,
  ip_col = "IPAddress",
  country = "US",
  print_tibble = TRUE,
  quiet = FALSE
)

Arguments

.data
Data frame or tibble (preferably exported from Qualtrics).

ip_col
Column name for IP addresses.

country
Two-letter abbreviation of country to check (default is "US").

print_tibble
Logical indicating whether to print returned tibble to console.

quiet
Logical indicating whether to print message to console.

Value

An object of the same type as .data that includes the rows with IP addresses outside of the specified country. For a function that marks these rows, use mark_ip(). For a function that excludes these rows, use exclude_ip().

Details

Default column names are set based on output from the qualtRics::fetch_survey(). The function uses iptools::country_ranges() to assign IP addresses to specific countries using ISO 3166-1 alpha-2 country codes.

The function outputs to console a message about the number of rows with IP addresses outside of the specified country. If there are NAs for IP addresses (likely due to including preview data---see check_preview()), it will print a message alerting to the number of rows with NAs.

Note

This function requires internet connectivity as it uses the iptools::country_ranges() function, which pulls daily updated data from http://www.iwik.org/ipcountry/.

See also

Examples

# Check for IP addresses outside of the US
data(qualtrics_text)
check_ip(qualtrics_text)

#> 2 out of 100 rows have NA values for IP addresses (likely because it includes preview data).

#> 4 out of 100 rows have IP addresses outside of US.

#> # A tibble: 4 x 16
#> # Rowwise: 
#>   StartDate           EndDate             Status IPAddress Progress
#>   <dttm>              <dttm>              <chr>  <chr>        <dbl>
#> 1 2020-12-11 16:59:08 2020-12-11 17:02:05 IP Ad… 84.56.18…      100
#> 2 2020-12-11 17:09:18 2020-12-11 17:16:57 IP Ad… 31.252.2…      100
#> 3 2020-12-17 15:47:37 2020-12-17 15:51:33 IP Ad… 89.167.1…      100
#> 4 2020-12-17 15:49:28 2020-12-17 15:55:06 IP Ad… 2.246.67…       44
#> # … with 11 more variables: `Duration (in seconds)` <dbl>, Finished <lgl>,
#> #   RecordedDate <dttm>, ResponseId <chr>, LocationLatitude <dbl>,
#> #   LocationLongitude <dbl>, UserLanguage <chr>, Browser <chr>, Version <chr>,
#> #   `Operating System` <chr>, Resolution <chr>

# Remove preview data first
qualtrics_text %>%
  exclude_preview() %>%
  check_ip()

#> 2 out of 100 preview rows were excluded, leaving 98 rows.

#> 4 out of 98 rows have IP addresses outside of US.

#> # A tibble: 4 x 16
#> # Rowwise: 
#>   StartDate           EndDate             Status IPAddress Progress
#>   <dttm>              <dttm>              <chr>  <chr>        <dbl>
#> 1 2020-12-11 16:59:08 2020-12-11 17:02:05 IP Ad… 84.56.18…      100
#> 2 2020-12-11 17:09:18 2020-12-11 17:16:57 IP Ad… 31.252.2…      100
#> 3 2020-12-17 15:47:37 2020-12-17 15:51:33 IP Ad… 89.167.1…      100
#> 4 2020-12-17 15:49:28 2020-12-17 15:55:06 IP Ad… 2.246.67…       44
#> # … with 11 more variables: `Duration (in seconds)` <dbl>, Finished <lgl>,
#> #   RecordedDate <dttm>, ResponseId <chr>, LocationLatitude <dbl>,
#> #   LocationLongitude <dbl>, UserLanguage <chr>, Browser <chr>, Version <chr>,
#> #   `Operating System` <chr>, Resolution <chr>

# Check for IP addresses outside of Germany
qualtrics_text %>%
  exclude_preview() %>%
  check_ip(country = "DE")

#> 2 out of 100 preview rows were excluded, leaving 98 rows.

#> 94 out of 98 rows have IP addresses outside of DE.

#> # A tibble: 94 x 16
#> # Rowwise: 
#>    StartDate           EndDate             Status IPAddress Progress
#>    <dttm>              <dttm>              <chr>  <chr>        <dbl>
#>  1 2020-12-11 12:17:22 2020-12-11 12:21:41 IP Ad… 73.23.43…      100
#>  2 2020-12-11 12:17:41 2020-12-11 12:22:07 IP Ad… 16.140.1…      100
#>  3 2020-12-11 12:19:45 2020-12-11 12:23:16 IP Ad… 107.57.2…      100
#>  4 2020-12-11 12:37:51 2020-12-11 12:43:09 IP Ad… 15.232.1…      100
#>  5 2020-12-11 12:41:23 2020-12-11 12:44:37 IP Ad… 24.195.9…      100
#>  6 2020-12-11 12:37:04 2020-12-11 12:45:50 IP Ad… 98.75.20…      100
#>  7 2020-12-11 12:55:35 2020-12-11 12:59:24 IP Ad… 51.113.1…      100
#>  8 2020-12-11 13:22:34 2020-12-11 13:35:19 IP Ad… 17.163.1…      100
#>  9 2020-12-11 17:02:00 2020-12-11 17:03:30 IP Ad… 70.203.6…      100
#> 10 2020-12-11 17:01:32 2020-12-11 17:09:41 IP Ad… 33.185.8…      100
#> # … with 84 more rows, and 11 more variables: `Duration (in seconds)` <dbl>,
#> #   Finished <lgl>, RecordedDate <dttm>, ResponseId <chr>,
#> #   LocationLatitude <dbl>, LocationLongitude <dbl>, UserLanguage <chr>,
#> #   Browser <chr>, Version <chr>, `Operating System` <chr>, Resolution <chr>

# Do not print rows to console
qualtrics_text %>%
  exclude_preview() %>%
  check_ip(print_tibble = FALSE)

#> 2 out of 100 preview rows were excluded, leaving 98 rows.

#> 4 out of 98 rows have IP addresses outside of US.

# Do not print message to console
qualtrics_text %>%
  exclude_preview() %>%
  check_ip(quiet = TRUE)

#> 2 out of 100 preview rows were excluded, leaving 98 rows.

#> # A tibble: 4 x 16
#> # Rowwise: 
#>   StartDate           EndDate             Status IPAddress Progress
#>   <dttm>              <dttm>              <chr>  <chr>        <dbl>
#> 1 2020-12-11 16:59:08 2020-12-11 17:02:05 IP Ad… 84.56.18…      100
#> 2 2020-12-11 17:09:18 2020-12-11 17:16:57 IP Ad… 31.252.2…      100
#> 3 2020-12-17 15:47:37 2020-12-17 15:51:33 IP Ad… 89.167.1…      100
#> 4 2020-12-17 15:49:28 2020-12-17 15:55:06 IP Ad… 2.246.67…       44
#> # … with 11 more variables: `Duration (in seconds)` <dbl>, Finished <lgl>,
#> #   RecordedDate <dttm>, ResponseId <chr>, LocationLatitude <dbl>,
#> #   LocationLongitude <dbl>, UserLanguage <chr>, Browser <chr>, Version <chr>,
#> #   `Operating System` <chr>, Resolution <chr>