Check for IP addresses that come from a specified country.

The check_ip() function subsets rows of data, retaining rows that have IP addresses in the specified country. The function is written to work with data from Qualtrics surveys.

check_ip(
  x,
  ip_col = "IPAddress",
  country = "US",
  print_tibble = TRUE,
  quiet = FALSE
)

Arguments

x
Data frame or tibble (preferably exported from Qualtrics).

ip_col
Column name for IP addresses.

country
Two-letter abbreviation of country to check (default is "US").

print_tibble
Logical indicating whether to print returned tibble to console.

quiet
Logical indicating whether to print message to console.

Value

An object of the same type as x that includes the rows with IP addresses outside of the specified country. For a function that marks these rows, use mark_ip(). For a function that excludes these rows, use exclude_ip().

Details

Default column names are set based on output from the qualtRics::fetch_survey(). The function uses iptools::country_ranges() to assign IP addresses to specific countries using ISO 3166-1 alpha-2 country codes.

The function outputs to console a message about the number of rows with IP addresses outside of the specified country. If there are NAs for IP addresses (likely due to including preview data---see check_preview()), it will print a message alerting to the number of rows with NAs.

Note

This function requires internet connectivity as it uses the iptools::country_ranges() function, which pulls daily updated data from http://www.iwik.org/ipcountry/.

See also

Examples

# Check for IP addresses outside of the US
data(qualtrics_text)
check_ip(qualtrics_text)

#> 2 out of 100 rows have NA values for IP addresses (likely because it includes preview data).

#> 4 out of 100 rows have IP addresses outside of US.

#>             StartDate             EndDate     Status   IPAddress Progress
#> 1 2020-12-11 16:59:08 2020-12-11 17:02:05 IP Address 84.56.189.0      100
#> 2 2020-12-11 17:09:18 2020-12-11 17:16:57 IP Address 31.252.29.0      100
#> 3 2020-12-17 15:47:37 2020-12-17 15:51:33 IP Address 89.167.14.0      100
#> 4 2020-12-17 15:49:28 2020-12-17 15:55:06 IP Address  2.246.67.0       44
#>   Duration (in seconds) Finished        RecordedDate        ResponseId
#> 1                    54     TRUE 2020-12-11 17:02:05 R_2RQ5kfCKKHudpj3
#> 2                   229     TRUE 2020-12-11 17:16:57 R_XUU5qhep0Xe1GyW
#> 3                   297     TRUE 2020-12-17 15:51:33 R_bEHQBqkV97IcXfD
#> 4                   355    FALSE 2020-12-17 15:55:06 R_UfSQq1VXYkVcxdJ
#>   LocationLatitude LocationLongitude UserLanguage Browser      Version
#> 1         43.23353         -77.55959           EN  Chrome 87.0.4280.88
#> 2         33.87347        -117.92215           EN  Chrome 87.0.4280.88
#> 3         35.39816         -78.81702           EN  Chrome 87.0.4280.88
#> 4         37.79058        -121.96737           EN  Chrome 87.0.4280.67
#>   Operating System Resolution
#> 1  Windows NT 10.0   1600x900
#> 2  Windows NT 10.0   1366x768
#> 3        Macintosh   1280x800
#> 4        Macintosh   1280x800

# Remove preview data first
qualtrics_text %>%
  exclude_preview() %>%
  check_ip()

#> 2 out of 100 preview rows were excluded, leaving 98 rows.

#> 0 out of 98 rows have IP addresses outside of US.

#>  [1] StartDate             EndDate               Status               
#>  [4] IPAddress             Progress              Duration (in seconds)
#>  [7] Finished              RecordedDate          ResponseId           
#> [10] LocationLatitude      LocationLongitude     UserLanguage         
#> [13] Browser               Version               Operating System     
#> [16] Resolution           
#> <0 rows> (or 0-length row.names)

# Check for IP addresses outside of Germany
qualtrics_text %>%
  exclude_preview() %>%
  check_ip(country = "DE")

#> 2 out of 100 preview rows were excluded, leaving 98 rows.

#> 0 out of 98 rows have IP addresses outside of DE.

#>  [1] StartDate             EndDate               Status               
#>  [4] IPAddress             Progress              Duration (in seconds)
#>  [7] Finished              RecordedDate          ResponseId           
#> [10] LocationLatitude      LocationLongitude     UserLanguage         
#> [13] Browser               Version               Operating System     
#> [16] Resolution           
#> <0 rows> (or 0-length row.names)

# Do not print rows to console
qualtrics_text %>%
  exclude_preview() %>%
  check_ip(print_tibble = FALSE)

#> 2 out of 100 preview rows were excluded, leaving 98 rows.

#> 0 out of 98 rows have IP addresses outside of US.

# Do not print message to console
qualtrics_text %>%
  exclude_preview() %>%
  check_ip(quiet = TRUE)

#> 2 out of 100 preview rows were excluded, leaving 98 rows.

#>  [1] StartDate             EndDate               Status               
#>  [4] IPAddress             Progress              Duration (in seconds)
#>  [7] Finished              RecordedDate          ResponseId           
#> [10] LocationLatitude      LocationLongitude     UserLanguage         
#> [13] Browser               Version               Operating System     
#> [16] Resolution           
#> <0 rows> (or 0-length row.names)